Add example react component to right sidebar
