+++ src/component.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export default function Comments() {
+  return <div>comments</div>;
+}
